# sms_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: sms-api
  version: 0.2.20-dev
paths:
  /analyses:
    post:
      tags:
      - Analysis - vEcoli
      summary: Run an analysis workflow (like multigeneration)
      operationId: run-analysis
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/AnalysisConfig'
              - type: 'null'
              title: Config
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJob'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Analysis - vEcoli
      summary: Get all available analysis output ids
      operationId: get-available-analyses
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                title: Response Get-Available-Analyses
    put:
      tags:
      - Analysis - vEcoli
      summary: Upload custom analysis module
      operationId: upload_analysis_module_analyses_put
      parameters:
      - name: submodule_name
        in: query
        required: true
        schema:
          type: string
          description: Submodule name(single, multiseed, etc)
          title: Submodule Name
        description: Submodule name(single, multiseed, etc)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_analysis_module_analyses_put'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Upload Analysis Module Analyses Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analyses/{id}:
    get:
      tags:
      - Analysis - vEcoli
      summary: Get an array of HTML files as strings and job status6
      operationId: get_analysis_analyses__id__get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                title: Response Get Analysis Analyses  Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /analyses/{id}/download:
    get:
      tags:
      - Analysis - vEcoli
      summary: Download a single file that was generated from a simulation analysis
        module
      operationId: download-analysis-output
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          title: Id
      - name: variant_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Variant Id
      - name: lineage_seed_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Lineage Seed Id
      - name: generation_id
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Generation Id
      - name: agent_id
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Agent Id
      - name: filename
        in: query
        required: true
        schema:
          type: string
          examples:
          - mass_fraction_summary.html
          title: Filename
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnalysisConfig:
      properties:
        analysis_options:
          $ref: '#/components/schemas/AnalysisConfigOptions'
        emitter_arg:
          additionalProperties:
            type: string
          type: object
          title: Emitter Arg
          default:
            out_dir: ''
      type: object
      required:
      - analysis_options
      title: AnalysisConfig
    AnalysisConfigOptions:
      properties:
        experiment_id:
          items:
            type: string
          type: array
          title: Experiment Id
        variant_data_dir:
          items:
            type: string
          type: array
          title: Variant Data Dir
        validation_data_path:
          items:
            type: string
          type: array
          title: Validation Data Path
        outdir:
          type: string
          title: Outdir
        cpus:
          type: integer
          title: Cpus
        single:
          additionalProperties: true
          type: object
          title: Single
        multidaughter:
          additionalProperties: true
          type: object
          title: Multidaughter
        multigeneration:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
          title: Multigeneration
          default:
            replication: {}
            ribosome_components: {}
            ribosome_crowding: {}
            ribosome_production: {}
            ribosome_usage: {}
            rna_decay_03_high: {}
        multiseed:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
          title: Multiseed
          default:
            protein_counts_validation: {}
            ribosome_spacing: {}
            subgenerational_expression_table: {}
        multivariant:
          additionalProperties:
            additionalProperties: true
            type: object
          type: object
          title: Multivariant
          default:
            average_monomer_counts: {}
            cell_mass: {}
            doubling_time_hist:
              skip_n_gens: 1
            doubling_time_line: {}
        multiexperiment:
          additionalProperties: true
          type: object
          title: Multiexperiment
          default: {}
      type: object
      required:
      - experiment_id
      - variant_data_dir
      - validation_data_path
      - outdir
      - cpus
      title: AnalysisConfigOptions
    AnalysisJob:
      properties:
        id:
          type: integer
          title: Id
        status:
          type: string
          title: Status
          default: WAITING
      type: object
      required:
      - id
      title: AnalysisJob
...
    Body_upload_analysis_module_analyses_put:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_analysis_module_analyses_put
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError